AutoGPT: extract `core.prompting` module out of `core.planning`
