+++ autogpts/autogpt/autogpt/core/planning/__init__.py
@@ -1,7 +1,5 @@
 """The planning system organizes the Agent's activities."""
 from autogpt.core.planning.schema import (
-    LanguageModelClassification,
-    LanguageModelPrompt,
     Task,
     TaskStatus,
     TaskType,
