+++ autogpts/autogpt/autogpt/core/planning/base.py
@@ -1,11 +1,3 @@
-import abc
-
-from autogpt.core.configuration import SystemConfiguration
-from autogpt.core.planning.schema import (
-    LanguageModelClassification,
-    LanguageModelPrompt,
-)
-
 # class Planner(abc.ABC):
 #     """Manages the agent's planning and goal-setting by constructing language model prompts."""
 #
@@ -57,20 +49,3 @@ from autogpt.core.planning.schema import (
 #
 #         """
 #         ...
-
-
-class PromptStrategy(abc.ABC):
-    default_configuration: SystemConfiguration
-
-    @property
-    @abc.abstractmethod
-    def model_classification(self) -> LanguageModelClassification:
-        ...
-
-    @abc.abstractmethod
-    def build_prompt(self, *_, **kwargs) -> LanguageModelPrompt:
-        ...
-
-    @abc.abstractmethod
-    def parse_response_content(self, response_content: dict) -> dict:
-        ...
