+++ autogpts/autogpt/autogpt/core/planning/schema.py
@@ -4,34 +4,6 @@ from typing import Optional
 from pydantic import BaseModel, Field
 
 from autogpt.core.ability.schema import AbilityResult
-from autogpt.core.resource.model_providers.schema import (
-    LanguageModelFunction,
-    LanguageModelMessage,
-)
-
-
-class LanguageModelClassification(str, enum.Enum):
-    """The LanguageModelClassification is a functional description of the model.
-
-    This is used to determine what kind of model to use for a given prompt.
-    Sometimes we prefer a faster or cheaper model to accomplish a task when
-    possible.
-
-    """
-
-    FAST_MODEL = "fast_model"
-    SMART_MODEL = "smart_model"
-
-
-class LanguageModelPrompt(BaseModel):
-    messages: list[LanguageModelMessage]
-    functions: list[LanguageModelFunction] = Field(default_factory=list)
-
-    def __str__(self):
-        return "\n\n".join(
-            f"{m.role.value.upper()}: {m.content}"
-            for m in self.messages
-        )
 
 
 class TaskType(str, enum.Enum):
