--- autogpts/autogpt/autogpt/core/planning/strategies/__init__.py
@@ -1,12 +1,0 @@
-from autogpt.core.planning.strategies.initial_plan import (
-    InitialPlan,
-    InitialPlanConfiguration,
-)
-from autogpt.core.planning.strategies.name_and_goals import (
-    NameAndGoals,
-    NameAndGoalsConfiguration,
-)
-from autogpt.core.planning.strategies.next_ability import (
-    NextAbility,
-    NextAbilityConfiguration,
-)
