+++ autogpts/autogpt/autogpt/core/prompting/__init__.py
@@ -1,0 +1,8 @@
+from .base import PromptStrategy
+from .schema import LanguageModelClassification, LanguageModelPrompt
+
+__all__ = [
+    "LanguageModelClassification",
+    "LanguageModelPrompt",
+    "PromptStrategy",
+]
