+++ autogpts/autogpt/autogpt/core/prompting/utils.py
